Add download.licham.io and dynamic DNS
This is a second nginx instance which hosts a directory on the server
for us to serve static public downloads from. This nginx container is
routed to via the nginx frontend and gets let's encrypt certs like
the other one.

Also add dynamic DNS for consistency with the rest of our services.